requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
chardet==3.0.4
discord.py==0.16.12
multidict==3.3.2
mysql-connector-python==8.0.5
websockets==3.4
